Refactor to prep for a new binary.
- Pulls out some functionality from `build_xray_quadtree.rs` into
  `generation.rs`. This is in preparation for adding a new binary that
  generates a single X-Ray image from a bounding box of the octree.
- Change `xray_from_points` to operate over non-square images.
